Factor out I2C functions
